requirements.txt
discord.py==2.0.1
discord_py_slash_command==4.2.1
nextcord==2.2.0
numpy==1.22.3
requests==2.27.1
yt_dlp==2022.5.18
